+++ flutter/lib/src/navigation/sentry_navigator_observer.dart
@@ -8,6 +8,7 @@ import 'package:meta/meta.dart';
 import '../native/native_frames.dart';
 import '../native/sentry_native_binding.dart';
 import 'time_to_display_tracker.dart';
+import 'time_to_full_display_tracker.dart';
 
 import '../../sentry_flutter.dart';
 import '../event_processor/flutter_enricher_event_processor.dart';
@@ -309,13 +310,18 @@ class SentryNavigatorObserver extends RouteObserver<PageRoute<dynamic>> {
       // Cancel unfinished TTID/TTFD spans, e.g this might happen if the user navigates
       // away from the current route before TTFD or TTID is finished.
       for (final child in (transaction as SentryTracer).children) {
+        if (child.finished) continue;
+
         final isTTIDSpan = child.context.operation ==
             SentrySpanOperations.uiTimeToInitialDisplay;
         final isTTFDSpan =
             child.context.operation == SentrySpanOperations.uiTimeToFullDisplay;
-        if (!child.finished && (isTTIDSpan || isTTFDSpan)) {
+        if (isTTIDSpan || isTTFDSpan) {
+          final finishTimestamp = isTTFDSpan
+              ? (ttidEndTimestampProvider() ?? endTimestamp)
+              : endTimestamp;
           await child.finish(
-            endTimestamp: endTimestamp,
+            endTimestamp: finishTimestamp,
             status: SpanStatus.deadlineExceeded(),
           );
         }